# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "图书馆管理系统API文档",
        "title": "Library Management System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/borrows": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据条件搜索借阅记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "借阅管理"
                ],
                "summary": "获取借阅列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "name": "book_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2024-01-01T00:00:00+08:00",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "minLength": 1,
                        "type": "string",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-01-01T00:00:00+08:00",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            1
                        ],
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户借阅图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "借阅管理"
                ],
                "summary": "借阅图书",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "借阅信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateBorrowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/borrows/return": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户归还图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "借阅管理"
                ],
                "summary": "归还图书",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "归还信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ReturnBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/borrows/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取指定借阅记录的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "借阅管理"
                ],
                "summary": "获取借阅详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "借阅ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员更新借阅记录信息 4-已取消 1-借阅中 2-已归还 3-已逾期",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "借阅管理"
                ],
                "summary": "更新借阅信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "借阅ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "借阅信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateBorrowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateBorrowRequest": {
            "type": "object",
            "required": [
                "book_id",
                "due_date"
            ],
            "properties": {
                "book_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "due_date": {
                    "description": "应还日期必须大于借阅日期",
                    "type": "string",
                    "example": "2024-01-01T00:00:00+08:00"
                },
                "remark": {
                    "type": "string",
                    "maxLength": 256,
                    "example": "请尽快归还"
                }
            }
        },
...
        "request.ReturnBookRequest": {
            "type": "object",
            "required": [
                "borrow_id"
            ],
            "properties": {
                "borrow_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "fine": {
                    "type": "number",
                    "minimum": 0,
                    "example": 10
                },
                "remark": {
                    "type": "string",
                    "maxLength": 256,
                    "example": "请尽快归还"
                }
            }
        },
...
        "request.UpdateBorrowRequest": {
            "type": "object",
            "required": [
                "due_date",
                "status"
            ],
            "properties": {
                "due_date": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00+08:00"
                },
                "fine": {
                    "type": "number",
                    "minimum": 0,
                    "example": 10
                },
                "remark": {
                    "type": "string",
                    "maxLength": 256,
                    "example": "请尽快归还"
                },
                "status": {
                    "description": "0-已取消 1-借阅中 2-已归还 3-已逾期",
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        3
                    ],
                    "example": 1
                }
            }
        },
...
        "response.Response": {
            "description": "API通用响应结构",
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}